test: reorganize frontend test layout
Move the frontend Vitest files into src/test while preserving the feature-based folder structure. Point Vitest at the dedicated test tree so setup and test discovery stay consistent.
