+++ frontend/src/test/App.test.tsx
@@ -1,28 +1,28 @@
 import { describe, it, expect, vi } from 'vitest'
 import { render, waitFor } from '@testing-library/react'
 import { MemoryRouter } from 'react-router-dom'
-import App from './App'
+import App from '../App'
 
 const getAdminRouteConfigMock = vi.hoisted(() => vi.fn())
 
-vi.mock('./admin/paths', async () => {
-  const actual = await vi.importActual<typeof import('./admin/paths')>('./admin/paths')
+vi.mock('../admin/paths', async () => {
+  const actual = await vi.importActual<typeof import('../admin/paths')>('../admin/paths')
   return {
     ...actual,
     getAdminRouteConfig: getAdminRouteConfigMock,
   }
 })
 
-vi.mock('./pages/admin/LoginPage', () => ({
+vi.mock('../pages/admin/LoginPage', () => ({
   default: () => <div>Admin login</div>,
   consumePrimedAdminSession: () => null,
 }))
 
-vi.mock('./pages/admin/DashboardPage', () => ({
+vi.mock('../pages/admin/DashboardPage', () => ({
   default: () => <div>Admin dashboard</div>,
 }))
 
-vi.mock('./pages/admin/ArchiveEditPage', () => ({
+vi.mock('../pages/admin/ArchiveEditPage', () => ({
   default: () => <div>Archive edit</div>,
 }))
