+++ frontend/src/test/auth/useAdminSession.test.tsx
@@ -1,6 +1,6 @@
 import { act, renderHook, waitFor } from '@testing-library/react'
 import { afterEach, describe, expect, it, vi } from 'vitest'
-import { useAdminSession } from './useAdminSession'
+import { useAdminSession } from '../../auth/useAdminSession'
 
 describe('useAdminSession', () => {
   afterEach(() => {
